🚚RENAME: change data set module place
**MODIFY:**
- 🧰 `Makefile`: change deprency place
- 🚚`src/main.c`: change include path

**RENAME:**
- 🚚`src/images/*` -> `src/data_set`: match module structure
