+++ src/main.c
@@ -10,7 +10,7 @@
 #include <stdio.h>
 #include "neural_network/neural_network.h"
 #include "neural_network/utils.h"
-#include "image/set.h"
+#include "data_set/set.h"
 
 #define nb_layers 4
 #define nb_test 10000
